refactor(agents): promote web_search contract from a dedicated output module
